refactor: divide App to small components
