+++ src/app/routes.tsx
@@ -1,0 +1,25 @@
+import React from 'react';
+import { AppRoutes } from '../constants/routes';
+import { About } from '../about/About';
+import { NotFound } from '../not-found/NotFound';
+import { ValidatedMainLayout } from './ValidatedMainLayout';
+
+export type RouteConfig = {
+  path: string;
+  element: React.ReactNode;
+};
+
+export const routes: RouteConfig[] = [
+  {
+    path: AppRoutes.ABOUT,
+    element: <About />,
+  },
+  {
+    path: AppRoutes.MAIN,
+    element: <ValidatedMainLayout />,
+  },
+  {
+    path: AppRoutes.NOT_FOUND,
+    element: <NotFound />,
+  },
+];
