+++ src/constants/routes.ts
@@ -1,0 +1,5 @@
+export enum AppRoutes {
+  ABOUT = '/about',
+  MAIN = '/:page?/:detailsId?',
+  NOT_FOUND = '*',
+}
