+++ src/main.tsx
@@ -2,7 +2,7 @@ import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
 import './index.css';
-import App from './app/App.tsx';
+import { App } from './app/App.tsx';
 import ErrorBoundary from './error-boundary/ErrorBoundary.js';
 
 const rootElement = document.getElementById('root');
