+++ src/not-found/NotFound.tsx
@@ -1,19 +1,15 @@
 import './NotFound.css';
 import { Link } from 'react-router-dom';
 
-function NotFound() {
-  return (
-    <div className="not-found-page">
-      <div className="not-found-content">
-        <h1>404</h1>
-        <h2>Page Not Found</h2>
-        <p>The page you are looking for does not exist.</p>
-        <Link to="/" className="home-link">
-          Go back to Home
-        </Link>
-      </div>
+export const NotFound = () => (
+  <div className="not-found-page">
+    <div className="not-found-content">
+      <h1>404</h1>
+      <h2>Page Not Found</h2>
+      <p>The page you are looking for does not exist.</p>
+      <Link to="/" className="home-link">
+        Go back to Home
+      </Link>
     </div>
-  );
-}
-
-export default NotFound;
+  </div>
+);
